Fix REDOS
Problem:
Many of the regexes are vulnerable to catastrophic backtracking.
A malicious userAgent string could lead to REDOS.

Solution:
Enforce maximum userAgent string length.

Effectiveness:
None of the regexes have an exponential blow-up.
The worst I observed would blow up with 2K characters.
This regex takes around a tenth of a second on a 1K attack string.
The others take on the order of milliseconds or faster.